Cleanup mailer by moving frequency check out and splitting into files
